failure strategies using runOn
This is the first subset of changes implementing failure strategies.
This PR implements `runOn` with two states `success` and `failure`:

task4:
    runOn:
        task: task1
        states: ["success"]
        task: task2
        states: ["failure"]
        task: task3
        states: ["success", "failure"]

Execute `task4` if:
- task1 was successful AND
- task2 was failure AND
- task3 was either successful OR failure

These changes should work for a pipeline with any number of branches i.e.

Pipeline with single branch:

Pipeline:
         task1
           | failure
         task2 (runOn task1 failure)
           | success
         task3 (runOn task2 success)

Pipeline with multiple branches:

Pipeline:
             task1
    success /     \ failure
         task2   task3

By default (without any runOn), Pipeline execution bails out on first task failure.
`runOn` confirms the state of the task specified matches with the actual state.
`runOn` Success is defined by TaskRunStatus with Condition.Type set to
`Succeeded` and Condition.Status set to `True`.
`runOn` Failure is defined by TaskRunStatus with Condition.Type set to
`Succeeded`, Condition.Status set to `False` and Condition.Reason set to `Failed`.

With `runOn`, pipelineRun attempts to execute all the specified tasks.
When a parent task does not match the state specified by its dependent task's
runOn section, dependent task is skipped and marked with SkippedAsStateConflicted
and continues executing rest of the tasks in pipeline. To clarify, in the above
example, if task1 succeeds, task2 is executed since it depends on task1's success
state and task3 is marked with SkippedAsStateConflicted but execution would continue
in case there are more tasks and pipeline ends successfully with `Completed`.

There are still two more states needs to be added as part of runOn, `skip` and
`always` as per the design document.