pipelinetask metadata
Adding metadata to PipelineTask to allow specifying metadata when a
task is embedded using taskSpec. This metadata will be propogated
to taskRun and then to the pods.

```
apiVersion: tekton.dev/v1beta1
kind: PipelineRun
metadata:
  name: pipelinerun-with-taskspec-to-echo-greetings
spec:
  pipelineSpec:
    tasks:
      - name: echo-greetings
        metadata:
          labels: [ …]
...
```

Metadata is already supported as part of Tasks and Pipelines while
respective CRDs are created. But was not possible to specify with
embedded resources.