declare a task is skipped
A task within pipelineState is considered skipped if one of the parents
has failed or conditionChecks failed.
GetPipelineConditionStatus builds a list of skipped tasks on every
reconcile cycle. We are iterating over the pipelineState to figure out if a task
is skipped if one of the parents had conditionChecks failed. Also
all the non started tasks are considered skipped when the pipeline is in
stopping mode. With finally tasks, this same condition does not apply. When
a dag task fails, we do want to skip rest of the dag tasks but not any
final tasks.

Instead of having this split logic in GetPipelineConditionStatus, declaring a
task skipped maintaining it in the pipelineRun.Status is the easiest and
simplest direction forward. Just like we recognize if a task has
failed/succeeded based on the taskRun object (not actual TaskRun), same could
be maintained and recognized for skipped tasks.