refactor param array indexing validation
This commit refactors param array indexing validation to extract all
references first, then validate if references are out of bound.
Previously we traverse all references and collect invalid references
there. The benefit of this refactoring is that it would be convenient
for other validation to reuse these references. This commit also moves
some shared functions into a separate pkg.

Signed-off-by: Yongxuan Zhang yongxuanzhang@google.com